RQADeforestation

Documentation for RQADeforestation.

RQADeforestation.agcube

RQADeforestation.anti_diagonal_density

RQADeforestation.countvalid

RQADeforestation.countvalid

RQADeforestation.countvalidint

RQADeforestation.gdalcube

RQADeforestation.grouptimes

RQADeforestation.inner_postprocessing

RQADeforestation.rqatrend

RQADeforestation.rqatrend

RQADeforestation.rqatrend

RQADeforestation.rqatrend_recurrenceanalysis

RQADeforestation.rqatrend_shuffle

RQADeforestation.smooth

RQADeforestation.agcube — Method

agcube(filenames) Open the underlying tiff files via ArchGDAL. This opens all files and keeps them open. This has a higher upfront cost, but might lead to a speedup down the line when we access the data.

RQADeforestation.anti_diagonal_density — Function

anti_diagonal_density(ts, thresh, metric)

Compute the average density of the diagonals perpendicular to the main diagonal for data series ts. Uses the threshold thresh and metric for the computation of the similarities.

RQADeforestation.countvalid — Method

countvalid(xout, xin)

Inner function to count the valid time steps in a datacube.
This function is aimed to be used inside of a mapCube call.

RQADeforestation.countvalid — Method

countvalid(cube)

Outer function to count the number of valid time steps in a cube.

RQADeforestation.countvalidint — Method

countvalidag(xout, xin)

Inner function to count the valid time steps in a datacube.
This function is aimed to be used inside of a mapCube call.

RQADeforestation.gdalcube — Method

gdalcube(indir, pol)

Load the datasets in indir with a polarisation pol as a ESDLArray. We assume, that indir is a folder with geotiffs in the same CRS which are mosaicked into timesteps and then stacked as a threedimensional array.

RQADeforestation.grouptimes — Function

grouptimes(times, timediff=200000) Group a sorted vector of time stamps into subgroups where the difference between neighbouring elements are less than timediff milliseconds. This returns the indices of the subgroups as a vector of vectors.

RQADeforestation.inner_postprocessing — Method

Compute the forest masking thresholding and clustering of the rqadata in one step

RQADeforestation.rqatrend — Function

rqatrend(xout, xin, thresh)

Compute the RQA trend metric for the non-missing time steps of xin, and save it to xout. thresh specifies the epsilon threshold of the Recurrence Plot computation

RQADeforestation.rqatrend — Method

rqatrend(path::AbstractString; thresh=2, outpath=tempname()*".zarr")

Compute the RQA trend metric for the data that is available on path.

RQADeforestation.rqatrend — Method

rqatrend(cube;thresh=2, path=tempname() * ".zarr")

Compute the RQA trend metric for the datacube cube with the epsilon threshold thresh.

RQADeforestation.rqatrend_recurrenceanalysis — Function

rqatrend(xout, xin, thresh)

Compute the RQA trend metric for the non-missing time steps of xin, and save it to xout. thresh specifies the epsilon threshold of the Recurrence Plot computation

RQADeforestation.rqatrend_shuffle — Method

rqatrend_shuffle(cube; thresh=2, path=tempname() * ".zarr", numshuffle=300)

Compute the RQA trend metric for shuffled time series of the data cube cube with the epsilon threshold thresh for numshuffle tries and save it into path.

RQADeforestation.smooth — Method

smooth(a, b, γ)

Weighted average of a and b with weight γ.

$(1 - γ) * a + γ * b$