RQADeforestation

Documentation for RQADeforestation.

RQADeforestation.classify_rqatrend

RQADeforestation.countvalid

RQADeforestation.countvalid

RQADeforestation.gdalcube

RQADeforestation.grouptimes

RQADeforestation.rqatrend

RQADeforestation.rqatrend

RQADeforestation.rqatrend

RQADeforestation.smooth

RQADeforestation.classify_rqatrend — Method

classify_rqatrend(trend; lowerbound=Float32(-5.0), upperbound=Float32(-0.5)))

Classify the rqatrend and put it into 254 bins so that they can fit into a UInt8 encoding. This is a compromise between data storage and accuracy of the change detection. The value range is 0 (no change) to 254 (definitive change) with 255 kept free as a Sentinel value for missing data.

RQADeforestation.countvalid — Method

countvalid(xout, xin)

Inner function to count the valid time steps in a datacube.
This function is aimed to be used inside of a mapCube call.

RQADeforestation.countvalid — Method

countvalid(cube)

Outer function to count the number of valid time steps in a cube.

RQADeforestation.gdalcube — Method

gdalcube(indir, pol)

Load the datasets in indir with a polarisation pol as a ESDLArray. We assume, that indir is a folder with geotiffs in the same CRS which are mosaicked into timesteps and then stacked as a threedimensional array.

RQADeforestation.grouptimes — Function

grouptimes(times, timediff=200000) Group a sorted vector of time stamps into subgroups where the difference between neighbouring elements are less than timediff milliseconds. This returns the indices of the subgroups as a vector of vectors.

RQADeforestation.rqatrend — Function

rqatrend(xout, xin, thresh)

Compute the RQA trend metric for the non-missing time steps of xin, and save it to xout. thresh specifies the epsilon threshold of the Recurrence Plot computation. lowerbound and upperbound are the bounds of the classification into UInt8. The result of rqatrend are UInt8 values between 0 (no change) to 254 (definitive change) with 255 as sentinel value for missing data.

RQADeforestation.rqatrend — Method

rqatrend(path::AbstractString; thresh=2, outpath=tempname()*".zarr")

Compute the RQA trend metric for the data that is available on path. See the rqatrend for a YAXArray for the description of the parameters.

RQADeforestation.rqatrend — Method

rqatrend(cube;thresh=2, path=tempname() * ".zarr")

Compute the RQA trend metric for the datacube cube with the epsilon threshold thresh. lowerbound and upperbound are forwarded to the classification of the RQA Trend result.

RQADeforestation.smooth — Method

smooth(a, b, γ)

Weighted average of a and b with weight γ.

$(1 - γ) * a + γ * b$